backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg[binary]==3.2.6
python-dotenv
SQLAlchemy==2.0.38
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
pydantic==2.10.6
email-validator==2.2.0
bcrypt==3.2.2
requests==2.31.0
celery==5.4.0
redis==5.0.1
